+++ editor/app/(demo)/demo/sales/campaign/polestar-kr-2503/portal/page.tsx
@@ -1,12 +1,12 @@
 "use client";
 import React from "react";
-import TMP_Portal from "./component";
+import Portal from "./_flows/page";
 import { ScreenWindowRoot } from "@/theme/templates/kit/components";
 
 export default function Home() {
   return (
     <ScreenWindowRoot>
-      <TMP_Portal />
+      <Portal />
     </ScreenWindowRoot>
   );
 }
