+++ editor/grida-react-canvas/viewport/ui/floating-bar.tsx
@@ -25,7 +25,7 @@ export function FloatingBar({
     <div
       className="group relative pointer-events-none"
       data-state={state}
-      data-is-component-consumer={isComponentConsumer}
+      data-layer-is-component-consumer={isComponentConsumer}
       style={data?.style}
     >
       {/* Title bar positioned above the parent using a percentage transform */}
@@ -50,7 +50,7 @@ export function FloatingBarContent({
     <div
       {...props}
       className={cn(
-        "max-w-full flex items-center gap-2 pointer-events-auto cursor-pointer rounded-lg py-2 px-2.5 truncate bg-background/80 group-data-[is-component-consumer=true]:bg-workbench-accent-violet/50 group-data-[state=hover]:bg-accent group-data-[state=active]:bg-accent",
+        "max-w-full flex items-center gap-2 pointer-events-auto cursor-pointer rounded-lg py-2 px-2.5 truncate bg-background/80 group-data-[layer-is-component-consumer='true']:bg-workbench-accent-violet/50 group-data-[state=hover]:bg-accent group-data-[state=active]:bg-accent",
         className
       )}
     >
@@ -68,7 +68,7 @@ export function FloatingBarTitle({
     <div
       {...props}
       className={cn(
-        "max-w-full w-min pointer-events-auto text-xs truncate text-muted-foreground/65 group-data-[is-component-consumer=true]:text-workbench-accent-violet group-data-[state=hover]:text-workbench-accent-sky group-data-[state=active]:text-workbench-accent-sky",
+        "max-w-full w-min pointer-events-auto text-xs truncate text-muted-foreground/65 group-data-[layer-is-component-consumer='true']:text-workbench-accent-violet group-data-[state=hover]:text-workbench-accent-sky group-data-[state=active]:text-workbench-accent-sky",
         className
       )}
     >
