+++ editor/scaffolds/editor/init.ts
@@ -448,10 +448,12 @@ function initialSiteEditorState(init: SiteDocumentEditorInit): EditorState {
                 name: "Invite Page",
                 type: "template_instance",
                 template_id: "tmp-2503-invite",
-                position: "relative",
+                position: "absolute",
                 removable: false,
                 active: true,
                 locked: false,
+                width: 375,
+                height: "auto",
                 properties: {},
                 props: {},
                 overrides: {},
@@ -461,10 +463,31 @@ function initialSiteEditorState(init: SiteDocumentEditorInit): EditorState {
                 name: "Join Page",
                 type: "template_instance",
                 template_id: "tmp-2503-join",
-                position: "relative",
+                position: "absolute",
                 removable: false,
                 active: true,
                 locked: false,
+                width: 375,
+                height: "auto",
+                properties: {},
+                props: {},
+                overrides: {},
+                top: 0,
+                left: 0,
+              },
+              join_hello: {
+                id: "join_hello",
+                name: "Join Hello (Overlay)",
+                type: "template_instance",
+                template_id: "tmp-2503-join-hello",
+                position: "absolute",
+                removable: false,
+                active: true,
+                locked: false,
+                width: 375,
+                height: 812,
+                top: 0,
+                left: -500,
                 properties: {},
                 props: {},
                 overrides: {},
@@ -474,13 +497,34 @@ function initialSiteEditorState(init: SiteDocumentEditorInit): EditorState {
                 name: "Portal Page",
                 type: "template_instance",
                 template_id: "tmp-2503-portal",
-                position: "relative",
+                position: "absolute",
                 removable: false,
                 active: true,
                 locked: false,
+                width: 375,
+                height: "auto",
                 properties: {},
                 props: {},
                 overrides: {},
+                top: 0,
+                left: 0,
+              },
+              portal_verify: {
+                id: "portal_verify",
+                name: "Verify (Overlay)",
+                type: "template_instance",
+                template_id: "tmp-2503-portal-verify",
+                position: "absolute",
+                removable: false,
+                active: true,
+                locked: false,
+                width: 375,
+                height: "auto",
+                properties: {},
+                props: {},
+                overrides: {},
+                top: 0,
+                left: 500,
               },
             },
             entry_scene_id: "invite",
@@ -492,7 +536,7 @@ function initialSiteEditorState(init: SiteDocumentEditorInit): EditorState {
                 children: ["invite"],
                 guides: [],
                 constraints: {
-                  children: "single",
+                  children: "multiple",
                 },
                 order: 1,
               },
@@ -500,10 +544,10 @@ function initialSiteEditorState(init: SiteDocumentEditorInit): EditorState {
                 type: "scene",
                 id: "join",
                 name: "Join",
-                children: ["join"],
+                children: ["join", "join_hello"],
                 guides: [],
                 constraints: {
-                  children: "single",
+                  children: "multiple",
                 },
                 order: 2,
               },
@@ -511,10 +555,10 @@ function initialSiteEditorState(init: SiteDocumentEditorInit): EditorState {
                 type: "scene",
                 id: "portal",
                 name: "Portal",
-                children: ["portal"],
+                children: ["portal", "portal_verify"],
                 guides: [],
                 constraints: {
-                  children: "single",
+                  children: "multiple",
                 },
                 order: 3,
               },
